backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
requests>=2.30.0
python-multipart==0.0.6
mediapipe==0.10.31
pandas==2.3.3
filterpy==1.4.5
numpy==2.2.6
matplotlib==3.10.8
openai
instructor
pydantic
python-dotenv
langchain
google-genai
opencv-python
numpy
supabase
groq